fix: ae content parsing hang

+++ Code/components/es_loader/ESLoader.h
@@ -45,6 +45,6 @@ private:
     fs::path m_loadOrderFile{};
     fs::path m_pluginsFile{};
     Vector<PluginData> m_loadOrder{};
-    TiltedPhoques::Map<String, uint8_t> m_masterFiles{};
+    TiltedPhoques::Map<String, uint32_t> m_masterFiles{};
 };
 } // namespace ESLoader
